+++ demisto_sdk/tests/integration_tests/generate_docs_integration_test.py
@@ -10,7 +10,7 @@ GENERATE_DOCS_CMD = "generate-docs"
 DEMISTO_SDK_PATH = join(git_path(), "demisto_sdk")
 
 
-class TestPlaybooks():
+class TestPlaybooks:
     def test_integration_generate_docs_playbook_positive_with_io(self, tmpdir):
         """
         Given
@@ -72,7 +72,6 @@ class TestPlaybooks():
         ]
         result = runner.invoke(main, arguments)
         readme_path = join(tmpdir, 'Playbooks.playbook-test_README.md')
-
         assert result.exit_code == 0
         assert 'Start generating playbook documentation...' in result.stdout
         assert not result.stderr
@@ -83,9 +82,77 @@ class TestPlaybooks():
             assert 'There are no inputs for this playbook.' in contents
             assert 'There are no outputs for this playbook.' in contents
 
+    def test_integration_generate_docs_playbook_dependencies_old_integration(self, tmpdir):
+        """
+        Given
+        - Path to valid playbook yml file to generate docs for.
+        - Path to directory to write the README.md file.
+
+        When
+        - Running the generate-docs command.
+
+        Then
+        - Ensure README.md is created.
+        - Ensure integration dependencies exists.
+        - Ensure Builtin not in dependencies.
+        """
+        valid_playbook_with_dependencies = join(DEMISTO_SDK_PATH, "tests/test_files/Packs/DummyPack/Playbooks/DummyPlaybook.yml")
+        runner = CliRunner(mix_stderr=False)
+        arguments = [
+            GENERATE_DOCS_CMD,
+            '-i', valid_playbook_with_dependencies,
+            '-o', tmpdir
+        ]
+        result = runner.invoke(main, arguments)
+        readme_path = join(tmpdir, 'DummyPlaybook_README.md')
+
+        assert result.exit_code == 0
+        assert 'Start generating playbook documentation...' in result.stdout
+        assert not result.stderr
+        assert not result.exception
+        assert Path(readme_path).exists()
+        with open(readme_path, 'r') as readme_file:
+            contents = readme_file.read()
+            assert 'Builtin' not in contents
+            assert '### Integrations\n* DummyIntegration\n' in contents
+
+    def test_integration_generate_docs_playbook_pack_dependencies(self, tmpdir):
+        """
+        Given
+        - Path to valid playbook yml file to generate docs for.
+        - Path to directory to write the README.md file.
+
+        When
+        - Running the generate-docs command.
+
+        Then
+        - Ensure README.md is created.
+        - Ensure integration dependencies exists.
+        - Ensure Builtin not in dependencies.
+        """
+        valid_playbook_with_dependencies = join(DEMISTO_SDK_PATH, "tests/test_files/Packs/CortexXDR/Playbooks/Cortex_XDR_Incident_Handling.yml")
+        runner = CliRunner(mix_stderr=False)
+        arguments = [
+            GENERATE_DOCS_CMD,
+            '-i', valid_playbook_with_dependencies,
+            '-o', tmpdir
+        ]
+        result = runner.invoke(main, arguments)
+        readme_path = join(tmpdir, 'Cortex_XDR_Incident_Handling_README.md')
+
+        assert result.exit_code == 0
+        assert 'Start generating playbook documentation...' in result.stdout
+        assert not result.stderr
+        assert not result.exception
+        assert Path(readme_path).exists()
+        with open(readme_path, 'r') as readme_file:
+            contents = readme_file.read()
+            assert 'Builtin' not in contents
+            assert '### Integrations\n* PaloAltoNetworks_XDR\n' in contents
+
 
 @pytest.mark.skip(reason='Just place-holder stubs for later implementation')
-class TestScripts():
+class TestScripts:
     def test_integration_generate_docs_script_positive(self):
         pass
 
@@ -94,7 +161,7 @@ class TestScripts():
 
 
 @pytest.mark.skip(reason='Just place-holder stubs for later implementation')
-class TestIntegrations():
+class TestIntegrations:
     def test_integration_generate_docs_integration_positive(self):
         pass